Move RemoveProduct dialog to own package
